+++ src/main/java/racingcar/RacingCarApplication.java
@@ -1,8 +1,7 @@
 package racingcar;
 
-import racingcar.service.Racing;
-import racingcar.service.RacingCar;
-import racingcar.service.RacingCars;
+import racingcar.domain.CarFactory;
+import racingcar.domain.RacingCar;
 import racingcar.view.InputView;
 
 import java.util.List;
@@ -16,13 +15,16 @@ public class RacingCarApplication {
         int numberOfCars = inputValues.get(0); // 자동차 수
         int numberOfTrials = inputValues.get(1); // 이동 시도 횟수
 
-        // 전체 자동차 객체 만들기
-        RacingCars racingCars = new RacingCars(numberOfCars);
+//        // 전체 자동차 객체 만들기
+//        RacingCars racingCars = new RacingCars(numberOfCars);
+//
+//        // 이동 시도 횟수에 따른 racing 진행
+//        Racing racing = new Racing(racingCars, numberOfTrials);
+//        racing.startRacing();
+//        RacingCars racingResult = racing.getRacing();
 
-        // 이동 시도 횟수에 따른 racing 진행
-        Racing racing = new Racing(racingCars, numberOfTrials);
-        racing.startRacing();
-        RacingCars racingResult = racing.getRacing();
+        List<RacingCar> racingCars = CarFactory.makeCars(numberOfCars);
+
 
     }
 }
