+++ src/main/java/racingcar/domain/RacingCar.java
@@ -1,10 +1,10 @@
-package racingcar.service;
+package racingcar.domain;
 
 public class RacingCar {
     final static int CAR_MOVEMENT_STANDARD_POINT = 4;   // random 값이 4이상인 경우 전진
     final static int DEFAULT_MOVEMENT_DISTANCE = 1;     // 자동차 이동 시 1만큼 전진
     private String carNumber;   // 차 넘버
-    private int position;       // 차가 이동한 위치
+    private int position = 1;       // 차가 이동한 위치
 
     public RacingCar(String carNumber) {
         this.carNumber = carNumber;
