+++ src/test/java/racingcar/RacingCarTest.java
@@ -2,7 +2,7 @@ package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.service.*;
+import racingcar.domain.*;
 
 
 import static org.assertj.core.api.Assertions.assertThat;
